fix error in RemoteBox in --verify mode
error was in Client.js RemoteBox and in Daemon.js frame.on()
when --verify was used.
client was not passing the second file with content being verified
to server side. On the server side frame.on() did not expect that client may
send a second file with content being verified.